Added value change completed signal for double spin boxes and double properties use this signal to avoid multiple value change notifications

+++ Libraries/OTWidgets/include/OTWidgets/PropertyInputDouble.h
@@ -33,7 +33,7 @@ namespace ot {
 		bool hasInputError(void) const;
 
 	private Q_SLOTS:
-		void lclValueChanged(int);
+		void lclValueChanged(double);
 		void lclTextChanged(void);
 		void lclEditingFinishedChanged(void);
